# backend/openapi.yaml
openapi: 3.0.3
info:
  title: FundooNotes API
  version: 1.0.0
  description: API for FundooNotes - Note taking and organization application
paths:
...
  /api/invitations/:
    get:
      operationId: invitations_retrieve
      description: Fetch all pending invitations for the authenticated user.
      summary: Get pending invitations for the current user
      tags:
      - invitations
      security:
      - bearerAuth: []
      - {}
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PendingInvitationsDocsEnvelope'
          description: ''
    post:
      operationId: invitations_create
      description: Accept or decline a pending invitation.
      summary: Respond to an invitation (accept or decline)
      tags:
      - invitations
      requestBody:
        content:
          type:
            schema:
              type: object
              additionalProperties: {}
              description: Unspecified request body
          properties:
            schema:
              invite_id:
                type: integer
              action:
                type: string
                enum:
                - accept
                - decline
          required:
            schema:
              type: object
              additionalProperties: {}
              description: Unspecified request body
      security:
      - bearerAuth: []
      - {}
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/InvitationActionDocsEnvelope'
          description: ''
        '400':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/InvitationActionDocsEnvelope'
          description: ''
...
  /api/notes/{note_id}/collaborators/:
    get:
      operationId: notes_collaborators_retrieve
      parameters:
      - in: path
        name: note_id
        schema:
          type: integer
        required: true
      tags:
      - notes
      security:
      - bearerAuth: []
      - {}
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CollaboratorDocsEnvelope'
          description: ''
    post:
      operationId: notes_collaborators_create
      parameters:
      - in: path
        name: note_id
        schema:
          type: integer
        required: true
      tags:
      - notes
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CollaboratorAdd'
          application/x-www-form-urlencoded:
            schema:
              $ref: '#/components/schemas/CollaboratorAdd'
          multipart/form-data:
            schema:
              $ref: '#/components/schemas/CollaboratorAdd'
        required: true
      security:
      - bearerAuth: []
      - {}
      responses:
        '201':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CollaboratorInviteDocsEnvelope'
          description: ''
  /api/notes/{note_id}/collaborators/{user_id}/:
    patch:
      operationId: notes_collaborators_partial_update
      parameters:
      - in: path
        name: note_id
        schema:
          type: integer
        required: true
      - in: path
        name: user_id
        schema:
          type: integer
        required: true
      tags:
      - notes
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PatchedCollaboratorAdd'
          application/x-www-form-urlencoded:
            schema:
              $ref: '#/components/schemas/PatchedCollaboratorAdd'
          multipart/form-data:
            schema:
              $ref: '#/components/schemas/PatchedCollaboratorAdd'
      security:
      - bearerAuth: []
      - {}
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CollaboratorDetailDocsEnvelope'
          description: ''
    delete:
      operationId: notes_collaborators_destroy
      parameters:
      - in: path
        name: note_id
        schema:
          type: integer
        required: true
      - in: path
        name: user_id
        schema:
          type: integer
        required: true
      tags:
      - notes
      security:
      - bearerAuth: []
      - {}
      responses:
        '204':
          description: No response body
...
components:
  schemas:
    CollaboratorAdd:
      type: object
      properties:
        email:
          type: string
          format: email
        access_level:
          enum:
          - view
          - edit
          type: string
          description: |-
            * `view` - view
            * `edit` - edit
          x-spec-enum-id: 0554c6b2e2490bc0
          default: view
      required:
      - email
    CollaboratorDetailDocsEnvelope:
      type: object
      properties:
        message:
          type: string
        payload:
          $ref: '#/components/schemas/CollaboratorDocs'
        status_code:
          type: integer
      required:
      - message
      - payload
      - status_code
    CollaboratorDocs:
      type: object
      properties:
        name:
          type: string
        email:
          type: string
          format: email
        access_level:
          type: string
      required:
      - access_level
      - email
      - name
    CollaboratorDocsEnvelope:
      type: object
      properties:
        message:
          type: string
        payload:
          type: array
          items:
            $ref: '#/components/schemas/CollaboratorDocs'
        status_code:
          type: integer
      required:
      - message
      - payload
      - status_code
    CollaboratorInviteDocs:
      type: object
      properties:
        name:
          type: string
        email:
          type: string
          format: email
        access_level:
          type: string
        status:
          type: string
        token:
          type: string
      required:
      - access_level
      - email
      - name
      - status
      - token
    CollaboratorInviteDocsEnvelope:
      type: object
      properties:
        message:
          type: string
        payload:
          $ref: '#/components/schemas/CollaboratorInviteDocs'
        status_code:
          type: integer
      required:
      - message
      - payload
      - status_code
    InvitationActionDocs:
      type: object
      properties:
        detail:
          type: string
      required:
      - detail
    InvitationActionDocsEnvelope:
      type: object
      properties:
        message:
          type: string
        payload:
          $ref: '#/components/schemas/InvitationActionDocs'
        status_code:
          type: integer
      required:
      - message
      - payload
      - status_code
...
    PatchedCollaboratorAdd:
      type: object
      properties:
        email:
          type: string
          format: email
        access_level:
          enum:
          - view
          - edit
          type: string
          description: |-
            * `view` - view
            * `edit` - edit
          x-spec-enum-id: 0554c6b2e2490bc0
          default: view
...
    PendingInvitationDocs:
      type: object
      properties:
        note_title:
          type: string
        invited_by_name:
          type: string
        invited_by_email:
          type: string
          format: email
        access_level:
          type: string
      required:
      - access_level
      - invited_by_email
      - invited_by_name
      - note_title
    PendingInvitationsDocsEnvelope:
      type: object
      properties:
        message:
          type: string
        payload:
          type: array
          items:
            $ref: '#/components/schemas/PendingInvitationDocs'
        status_code:
          type: integer
      required:
      - message
      - payload
      - status_code
...
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Enter your JWT token